+++ anknotes/find_deleted_notes.py
@@ -1,20 +1,24 @@
 # -*- coding: utf-8 -*-
-try:    from pysqlite2 import dbapi2 as sqlite
-except ImportError: from sqlite3 import dbapi2 as sqlite
+try:
+    from pysqlite2 import dbapi2 as sqlite
+except ImportError:
+    from sqlite3 import dbapi2 as sqlite
 
 from anknotes.shared import *
 
-Error = sqlite.Error 
+Error = sqlite.Error
 ankDBSetLocal()
 
 PATH = os.path.dirname(os.path.abspath(__file__)) + '\\extra\\testing\\'
-ENNotes = file(os.path.join(PATH, "^ Scratch Note.enex") , 'r').read()
+ENNotes = file(os.path.join(PATH, "^ Scratch Note.enex"), 'r').read()
 # find = file(os.path.join(PATH, "powergrep-find.txt") , 'r').read().splitlines()
 # replace = file(os.path.join(PATH, "powergrep-replace.txt") , 'r').read().replace('https://www.evernote.com/shard/s175/nl/19775535/' , '').splitlines()
 
 all_notes = ankDB().execute("SELECT guid, title FROM %s " % TABLES.EVERNOTE.NOTES)
 find_guids = {}
 
+
+# noinspection PyShadowingNames
 def clean_title(title):
     title = title.replace('&amp;', '&')
     title = title.replace('&apos;', "'")
@@ -23,10 +27,11 @@ def clean_title(title):
     title = title.replace('&amp;', '&')
     title = title.replace('&quot;', '"')
     title = title.replace('&nbsp;', ' ')
-    if isinstance(title , str):
-        title = unicode(title , 'utf-8')     
-    title = title.replace(u'\xa0', ' ')  
+    if isinstance(title, str):
+        title = unicode(title, 'utf-8')
+    title = title.replace(u'\xa0', ' ')
     return title
+
 
 for line in all_notes:
     # line = line.split('::: ')
@@ -34,31 +39,30 @@ for line in all_notes:
     # title = line[1]
     guid = line['guid']
     title = line['title']
-    title = clean_title(title)  
+    title = clean_title(title)
     find_guids[guid] = title
-    
 
-for match in find_evernote_links(ENNotes): 
+for match in find_evernote_links(ENNotes):
     guid = match.group('guid')
     title = match.group('Title')
-    title = clean_title(title)                 
+    title = clean_title(title)
     title_safe = str_safe(title)
     if guid in find_guids:
         find_title = find_guids[guid]
-        find_title_safe = str_safe(find_title)    
+        find_title_safe = str_safe(find_title)
         if find_title_safe == title_safe:
             del find_guids[guid]
         else:
-            print("Found guid match, title mismatch for %s: \n - %s\n - %s" % (guid, title_safe, find_title_safe)        )
+            print("Found guid match, title mismatch for %s: \n - %s\n - %s" % (guid, title_safe, find_title_safe))
     else:
         title_safe = str_safe(title)
-        print("COULD NOT FIND guid for %s: %s" % (guid, title_safe)    )
+        print("COULD NOT FIND guid for %s: %s" % (guid, title_safe))
 
-dels = []        
-with open(os.path.join(PATH, 'deleted-notes.txt') , 'w') as filesProgress:               
+dels = []
+with open(os.path.join(PATH, 'deleted-notes.txt'), 'w') as filesProgress:
     for guid, title in find_guids.items():
-        print>>filesProgress, str_safe('%s::: %s' % (guid, title))
+        print>> filesProgress, str_safe('%s::: %s' % (guid, title))
         dels.append([guid])
-        
+
 ankDB().executemany("DELETE FROM %s WHERE guid = ?" % TABLES.EVERNOTE.NOTES, dels)
 ankDB().commit()
